Remove useField hook and replace with `controlled` method
