+++ lib/index.ts
@@ -1,2 +1,2 @@
 export * from './useForm'
-export * from './useField'
+export * from './field'
